Allow whole issuer to be specified
This is contrary to using string concatenation to build up an URL.

The reason for this is that EG. token verification does a strict
comparison between supplied issuer and the iss property of the JWT. An
issuer however may not have an ending "/". An issuer may for instance
have an URL component part.

Allowing the whole issuer to be supplied gives the consumer more
flexibility. Right now, the library is somewhere between a) being
tightly coupled to Auth0 and b) only coupled to the OpenID interface.
This takes it somewhat closer to b). Being close to b) and removing
traces of Auth0 would open it up to a whole class of providers.

I've found some other things that ensures that we're still at a).

  1. Unsecurity expects to receive claim values from the ID Token, but I
  believe this is optional according to the spec [1]. To be certain of
  the availability, the userinfo endpoint should be requested.

  2. We're sending JSON-encoded requests to the token endpoint. This is
  not in accordance with the spec [2-3], albeit many providers allow it
  (and it so happens that Auth0 is one of them).

  3. The URL for the token endpoint is statically coded, but may take
  different form for different providers. To fully comply, one should
  perform a dicovery by requesting the well known configuration URL and
  use the returned information to make further requests.

[1] https://openid.net/specs/openid-connect-core-1_0.html#ScopeClaims
[2] https://openid.net/specs/openid-connect-core-1_0.html#TokenRequest
[3] https://tools.ietf.org/html/rfc6749#section-4.1.3