Include `data` when creating internal server errors
This appears to have been left out by accident and prevents the
application from forwarding the correct data associated with an internal
server error, which in turn makes it harder to debug applications.

An example of this can be found in Auth0OidcSecurityContext.scala:200,
where the `data` parameter is used to provide the actual error message.